desafio3/web/requirements.txt
Flask
redis
psycopg2-binary